Fix noise to be generated as float32. Improves display performance.

+++ nionswift_plugin/usim/Noise.py
@@ -13,5 +13,5 @@ class PoissonNoise:
     def apply(self, input: DataAndMetadata.DataAndMetadata) -> DataAndMetadata.DataAndMetadata:
         if self.enabled and self.poisson_level:
             rs = numpy.random.RandomState()  # use this to avoid blocking other calls to poisson
-            return input + (rs.poisson(self.poisson_level, size=input.data.shape) - self.poisson_level)
+            return input + (rs.poisson(self.poisson_level, size=input.data.shape).astype(input.data.dtype) - self.poisson_level)
         return input
